auto: claude changes

+++ backend/app/schemas/process.py
@@ -1,5 +1,4 @@
 from pydantic import BaseModel, Field
-from decimal import Decimal
 from typing import Optional
 
 
@@ -10,8 +9,8 @@ class ProcessResponse(BaseModel):
     name: str
     work_center: str = Field(..., alias="workCenter")
     standard_time: float = Field(..., alias="standardTime")
-    unit_price: Optional[Decimal] = Field(None, alias="unitPrice")
-    vave_price: Optional[Decimal] = Field(None, alias="vavePrice")
+    unit_price: Optional[float] = Field(None, alias="unitPrice")
+    vave_price: Optional[float] = Field(None, alias="vavePrice")
     has_history_data: bool = Field(..., alias="hasHistoryData")
 
     model_config = {"populate_by_name": True, "by_alias": True}
